# src/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PortPulse AI API",
    "description": "PortPulse AI: forecasting, anomalies, CP-SAT optimisation, routing, and IBM Bob agent orchestration.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/optimise": {
      "post": {
        "tags": [
          "optimise"
        ],
        "summary": "Run Optimise",
        "operationId": "run_optimise_api_optimise_post",
        "parameters": [
          {
            "name": "incremental",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "alias for body.incremental (W3)",
              "title": "Incremental"
            },
            "description": "alias for body.incremental (W3)"
          },
          {
            "name": "tidal",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "alias for body.tidal (W3)",
              "title": "Tidal"
            },
            "description": "alias for body.tidal (W3)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__optimise__ScenarioBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimiseLatest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tides": {
      "get": {
        "tags": [
          "optimise"
        ],
        "summary": "Get Tides",
        "description": "W3: tidal depth curve for every berth (read-only).\n\nServes whatever the tide pipeline last persisted — NOAA CO-OPS rows when present,\nelse the labelled harmonic model — as depth_ft at each integer hour 0..hours per\nberth. A GET never fetches or writes: the write side lives at startup (main.py)\nand POST /api/tides/refresh.",
        "operationId": "get_tides_api_tides_get",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 24,
              "default": 72,
              "title": "Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TidesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tides/refresh": {
      "post": {
        "tags": [
          "optimise"
        ],
        "summary": "Refresh Tides",
        "description": "Force a fresh NOAA CO-OPS fetch and replace tidal window rows.\n\nReturns {source, rows_written, station} — callable from the QualityPage refresh button.",
        "operationId": "refresh_tides_api_tides_refresh_post",
        "parameters": [
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 24,
              "default": 96,
              "title": "Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/plan": {
      "get": {
        "tags": [
          "plan"
        ],
        "summary": "Get Plan",
        "description": "Return the latest persisted plan or build a fresh engine-grounded plan.",
        "operationId": "get_plan_api_plan_get",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Text"
            }
          },
          {
            "name": "persist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Persist"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "plan"
        ],
        "summary": "Regenerate",
        "operationId": "regenerate_api_plan_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/export": {
      "get": {
        "tags": [
          "plan"
        ],
        "summary": "Export",
        "operationId": "export_api_export_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "assignments",
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OptimiseLatest": {
        "properties": {
          "run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "solver": {
            "type": "string",
            "title": "Solver"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "objective": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "solve_ms": {
            "type": "integer",
            "title": "Solve Ms"
          },
          "assignments": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Assignments"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "baseline": {
            "additionalProperties": true,
            "type": "object",
            "title": "Baseline"
          },
          "deltas": {
            "additionalProperties": true,
            "type": "object",
            "title": "Deltas"
          },
          "deferred": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Deferred"
          },
          "weights": {
            "additionalProperties": true,
            "type": "object",
            "title": "Weights"
          },
          "tidal_feasible": {
            "type": "boolean",
            "title": "Tidal Feasible",
            "default": true
          },
          "incremental": {
            "type": "boolean",
            "title": "Incremental",
            "default": false
          },
          "gap_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap Pct"
          },
          "horizon_hours": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon Hours"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "solver",
          "status",
          "solve_ms",
          "assignments",
          "metrics",
          "baseline",
          "deltas",
          "deferred",
          "weights"
        ],
        "title": "OptimiseLatest"
      },
...
      "PlanResponse": {
        "properties": {
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          },
          "shifts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Shifts"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "narrative_source": {
            "type": "string",
            "title": "Narrative Source"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "forecast_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forecast Run Id"
          },
          "optimiser_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optimiser Run Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "summary",
          "shifts",
          "text",
          "narrative_source"
        ],
        "title": "PlanResponse",
        "description": "Spec §6 traceability: the plan exposes the snapshot that produced it."
      },
...
      "TidesResponse": {
        "properties": {
          "period_hours": {
            "type": "number",
            "title": "Period Hours"
          },
          "amplitude_ft": {
            "type": "number",
            "title": "Amplitude Ft"
          },
          "under_keel_margin_ft": {
            "type": "number",
            "title": "Under Keel Margin Ft"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "berths": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Berths"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "period_hours",
          "amplitude_ft",
          "under_keel_margin_ft",
          "source",
          "berths"
        ],
        "title": "TidesResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__routers__optimise__ScenarioBody": {
        "properties": {
          "crane_factor": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.5,
            "title": "Crane Factor",
            "description": "crane availability multiplier",
            "default": 1.0
          },
          "move_rate_per_crane_hour": {
            "type": "number",
            "maximum": 35.0,
            "minimum": 20.0,
            "title": "Move Rate Per Crane Hour",
            "default": 28.0
          },
          "incremental": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incremental",
            "description": "warm-start from the previous CP-SAT solution (W3)"
          },
          "tidal": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tidal",
            "description": "enforce tidal windows on deep-draft vessels (W3)"
          }
        },
        "type": "object",
        "title": "ScenarioBody"
      },
...
    }
  }
}